requirements-dev.txt
# dev
pynvim
mypy
ipython
ipdb
pylint
black
pymongo-stubs
types-PyYAML
types-requests
